zalora_url_shortening/zalora_urlshortener/api/serializers.py
from rest_framework import serializers
from zalora_urlshortener.models import Shortener

class ShortenerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Shortener
        fields = "__all__"